refactor: 패키지 구조 변경 및 연관관계 삭제
