--- src/main/java/com/kernel/_SenseLog/common/dto/ResponseDTO.java
@@ -1,44 +1,0 @@
-package com.kernel._SenseLog.common.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.kernel._SenseLog.common.exception.BaseException;
-import lombok.*;
-
-import java.io.Serializable;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonInclude
-public class ResponseDTO<T> implements Serializable {
-
-    private ResultObject result;
-
-    private T data;
-
-    public ResponseDTO(ResultObject result) {
-        this.result = result;
-    }
-
-    public ResponseDTO(T data) {
-        this.data = data;
-    }
-
-    public static <T> ResponseDTO<T> ok() {
-        return new ResponseDTO<>(ResultObject.getSuccess());
-    }
-
-    public static <T> ResponseDTO<T> ok(T data) {
-        return new ResponseDTO<>(ResultObject.getSuccess(), data);
-    }
-
-    public static <T> ResponseDTO<T> response(T data) {
-        return new ResponseDTO<>(ResultObject.getSuccess(), data);
-    }
-
-    public ResponseDTO(BaseException ex) {
-        this.result = new ResultObject(ex);
-    }
-}
