--- src/main/java/com/kernel/_SenseLog/diaryLike/service/DiaryLikeService.java
@@ -1,7 +1,0 @@
-package com.kernel._SenseLog.diaryLike.service;
-
-import com.kernel._SenseLog.diaryLike.dto.DiaryLikeResponse;
-
-public interface DiaryLikeService {
-    DiaryLikeResponse toggleLike(Long userId, Long diaryId);
-}
