--- src/main/java/com/kernel/_SenseLog/user/entity/User.java
@@ -1,41 +1,0 @@
-package com.kernel._SenseLog.user.entity;
-
-import com.kernel._SenseLog.common.entity.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Getter
-@Table(name = "users")
-public class User extends BaseTimeEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(nullable = false, length = 255)
-  private String email;
-
-  @Column(nullable = false, length = 255)
-  private String password;
-
-  @Column(nullable = false, length = 50)
-  private String nickname;
-
-  @Builder
-  public User(String email, String password, String nickname) {
-    this.email = email;
-    this.password = password;
-    this.nickname = nickname;
-  }
-}
