+++ src/main/java/com/kernel/sense_log/common/exception/APIExceptionHandler.java
@@ -1,0 +1,21 @@
+package com.kernel.sense_log.common.exception;
+
+import com.kernel.sense_log.common.dto.ResponseDTO;
+import com.kernel.sense_log.common.dto.ResultObject;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class APIExceptionHandler extends RuntimeException {
+
+    @ExceptionHandler(BaseException.class)
+    protected ResponseDTO<ResultObject> handleBaseException(BaseException e,
+                                                            HttpServletRequest request, HttpServletResponse response) {
+        log.error(e.getMessage(), e);
+        return new ResponseDTO<>(e);
+    }
+}
