+++ src/main/java/com/kernel/sense_log/domain/entity/Diary.java
@@ -1,0 +1,36 @@
+package com.kernel.sense_log.domain.entity;
+
+import com.kernel.sense_log.common.entity.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor
+@Table(name = "diaries")
+public class Diary extends BaseTimeEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(nullable = false, length = 50)
+  private String content;
+
+  @Column(nullable = false)
+  private Boolean isPrivate;
+
+  @Column(nullable = false, length = 50)
+  private String aiMessage;
+
+  @Column(nullable = false)
+  private Long userId;
+
+
+}
