+++ src/main/java/com/kernel/sense_log/domain/entity/enumeration/emotion.java
@@ -1,0 +1,9 @@
+package com.kernel.sense_log.domain.entity.enumeration;
+
+public enum emotion {
+    매우_좋음,
+    좋은,
+    보통,
+    슬픔,
+    매우_슬픔;
+}
