+++ src/main/java/com/kernel/sense_log/domain/entity/enumeration/subEmotion.java
@@ -1,0 +1,24 @@
+package com.kernel.sense_log.domain.entity.enumeration;
+
+public enum subEmotion {
+    신나는,
+    편안한,
+    뿌듯한,
+    기대되는,
+    행복한,
+    의욕적인,
+    설레는,
+    상쾌한,
+    차분한,
+    감사한,
+    우울한,
+    외로운,
+    불안한,
+    슬픈,
+    화난,
+    부담되는,
+    짜증나는,
+    피곤한,
+    스트레스,
+    지루한;
+}
