--- src/test/java/com/kernel/_SenseLog/ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kernel._SenseLog;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
