+++ src/test/java/com/kernel/sense_log/ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.kernel.sense_log;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
